Move main method to TestServerApp class
