+++ servers/jak/webservice/app/build.gradle
@@ -234,7 +234,7 @@ jar {
     archiveAppendix = "WebService"
     archiveVersion = "${SERVER_VERSION}_${COUCHBASE_LITE_VERSION}"
 
-    manifest { attributes('Main-Class': 'com.couchbase.lite.javaws.mobiletest.EmbeddedJettyServer') }
+    manifest { attributes('Main-Class': 'com.couchbase.lite.javaws.mobiletest.TestServerApp') }
     from {
         configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
     } {
