+++ servers/jak/webservice/app/src/main/java/com/couchbase/lite/javaws/mobiletest/TestServerApp.java
@@ -17,6 +17,9 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.servlet.ServletContextHandler;
+
 import com.couchbase.lite.internal.utils.StringUtils;
 import com.couchbase.lite.mobiletest.GetDispatcher;
 import com.couchbase.lite.mobiletest.PostDispatcher;
@@ -42,6 +45,22 @@ public class TestServerApp extends HttpServlet {
 
     private transient GetDispatcher getDispatcher;
     private transient PostDispatcher postDispatcher;
+
+    public static void main(String[] args) throws Exception {
+        // Create a basic Jetty server instance
+        Server server = new Server(8080);
+
+        // Set up a servlet context and map servlets
+        ServletContextHandler context = new ServletContextHandler(ServletContextHandler.SESSIONS);
+        context.setContextPath("/");
+        server.setHandler(context);
+
+        context.addServlet(TestServerApp.class, "/");
+
+        // Start the server
+        server.start();
+        server.join();
+    }
 
     @SuppressFBWarnings("DM_DEFAULT_ENCODING")
     @Override
